refactor(core): remove old homescreen rendering from mercury UI
[no changelog]
